requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==19.3.0
boto3==1.14.46
boto3-stubs==1.14.46.0
botocore==1.17.46
cffi==1.14.2
chardet==3.0.4
discord==1.0.1
discord.py==1.4.1
docutils==0.15.2
idna==2.10
idna-ssl==1.1.0
jmespath==0.10.0
multidict==4.7.6
mypy-boto3==1.14.46.0
pycparser==2.20
pydub==0.24.1
PyNaCl==1.3.0
python-dateutil==2.8.1
python-decouple==3.6
PyYAML==5.4
s3transfer==0.3.3
six==1.15.0
tinydb==4.5.1
typing-extensions==3.7.4.2
urllib3==1.25.11
yarl==1.5.1
youtube-dl==2021.12.17
